requirements.txt
bottle==0.12.19
Pillow==9.1.0
